Rename Story to Steps
The Story section has been renamed to become the
Steps section.
